+++ Project/AlphaEngine/PhysicsEngine.cpp
@@ -6,7 +6,7 @@ void PhysicsEngine::Init(std::vector<Entity *> *entityList) {
 }
 
 void PhysicsEngine::Update(double dt) {
-	for(int i = 0; i < m_entities->size(); ++i) {
+	for(int i = 0; i < static_cast<int>(m_entities->size()); ++i) {
 		(*m_entities)[i]->GetCollider()->Update(m_entities, i, dt);
 	}
 }
